added a query file
