+++ rasa/actions/mysqlconn.py
@@ -11,14 +11,5 @@ def connect():
     return mydb
 
 
-def check_mail(email):
-    mydb = connect()
-    mycursor = mydb.cursor()
-    mycursor.execute("SELECT * FROM accounts_customer where email = '{}';".format(email))
-    myresult = mycursor.fetchall()
-    mydb.close()
-    if len(myresult):
-        return True
-    return False
