+++ rasa/actions/query.py
@@ -1,0 +1,11 @@
+from .mysqlconn import connect
+ 
+def check_mail(email):
+    mydb = connect()
+    mycursor = mydb.cursor()
+    mycursor.execute("SELECT * FROM accounts_customer where email = '{}';".format(email))
+    myresult = mycursor.fetchall()
+    mydb.close()
+    if len(myresult):
+        return True
+    return False
